LayoutTests/webaudio/oscillatornode-legacy-api.html

<!DOCTYPE html>
<html>
<head>
    <script src="../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("Testing that the legacy OscillatorNode API is not exposed on the unprefixed OscillatorNode");

        shouldBeUndefined("OscillatorNode.prototype.playbackState");

        // Tests backward compatibility for prefixed API.
        shouldBe("OscillatorNode.UNSCHEDULED_STATE", "0");
        shouldBe("OscillatorNode.SCHEDULED_STATE", "1");
        shouldBe("OscillatorNode.PLAYING_STATE", "2");
        shouldBe("OscillatorNode.FINISHED_STATE", "3");

        let legacyContext = new webkitAudioContext;
        let legacyOscillator = legacyContext.createOscillator();
        shouldBeDefined("legacyOscillator.playbackState");
        shouldBe("legacyOscillator.UNSCHEDULED_STATE", "0");
        shouldBe("legacyOscillator.SCHEDULED_STATE", "1");
        shouldBe("legacyOscillator.PLAYING_STATE", "2");
        shouldBe("legacyOscillator.FINISHED_STATE", "3");
    </script>
</body>
</html>
